Prevent selection of all nodes in single row mode when using new selection config API (and update tests)
