+++ testing/behavioural/src/selection/row-selection.test.ts
@@ -16,7 +16,7 @@ import {
     toggleHeaderCheckboxByIndex,
 } from './utils';
 
-describe('ag-grid Selection State', () => {
+describe('Row Selection Grid Options', () => {
     const columnDefs = [{ field: 'sport' }];
     const rowData = [
         { sport: 'football' },
@@ -71,7 +71,11 @@ describe('ag-grid Selection State', () => {
     describe('User Interactions', () => {
         describe('Single Row Selection', () => {
             test('Select single row', () => {
-                const api = createGrid({ columnDefs, rowData, rowSelection: 'single' });
+                const api = createGrid({
+                    columnDefs,
+                    rowData,
+                    selection: { mode: 'singleRow' },
+                });
 
                 clickRowByIndex(2);
 
@@ -79,7 +83,11 @@ describe('ag-grid Selection State', () => {
             });
 
             test('Clicking two rows selects only the last clicked row', () => {
-                const api = createGrid({ columnDefs, rowData, rowSelection: 'single' });
+                const api = createGrid({
+                    columnDefs,
+                    rowData,
+                    selection: { mode: 'singleRow' },
+                });
 
                 clickRowByIndex(2);
                 clickRowByIndex(5);
@@ -88,7 +96,11 @@ describe('ag-grid Selection State', () => {
             });
 
             test("SHIFT-click doesn't select multiple rows in single row selection mode", () => {
-                const api = createGrid({ columnDefs, rowData, rowSelection: 'single' });
+                const api = createGrid({
+                    columnDefs,
+                    rowData,
+                    selection: { mode: 'singleRow' },
+                });
 
                 clickRowByIndex(2);
                 clickRowByIndex(5, { shiftKey: true });
@@ -97,7 +109,11 @@ describe('ag-grid Selection State', () => {
             });
 
             test("CTRL-click doesn't select multiple rows in single row selection mode", () => {
-                const api = createGrid({ columnDefs, rowData, rowSelection: 'single' });
+                const api = createGrid({
+                    columnDefs,
+                    rowData,
+                    selection: { mode: 'singleRow' },
+                });
 
                 clickRowByIndex(2);
                 clickRowByIndex(5, { metaKey: true });
@@ -105,12 +121,14 @@ describe('ag-grid Selection State', () => {
                 assertSelectedRowsByIndex([5], api);
             });
 
-            test('suppressRowClickSelection prevents row from being selected when clicked', () => {
+            test('suppressClickSelection prevents row from being selected when clicked', () => {
                 const api = createGrid({
                     columnDefs,
                     rowData,
-                    rowSelection: 'single',
-                    suppressRowClickSelection: true,
+                    selection: {
+                        mode: 'singleRow',
+                        suppressClickSelection: true,
+                    },
                 });
 
                 clickRowByIndex(2);
@@ -122,8 +140,10 @@ describe('ag-grid Selection State', () => {
                 const api = createGrid({
                     columnDefs,
                     rowData,
-                    rowSelection: 'single',
-                    isRowSelectable: (node) => node.data.sport !== 'football',
+                    selection: {
+                        mode: 'singleRow',
+                        isRowSelectable: (node) => node.data.sport !== 'football',
+                    },
                 });
 
                 clickRowByIndex(0);
@@ -136,8 +156,7 @@ describe('ag-grid Selection State', () => {
                 const api = createGrid({
                     columnDefs,
                     rowData,
-                    rowSelection: 'multiple',
-                    isRowSelectable: (node) => node.data.sport !== 'football',
+                    selection: { mode: 'multiRow', isRowSelectable: (node) => node.data.sport !== 'football' },
                 });
 
                 clickRowByIndex(0);
@@ -155,7 +174,7 @@ describe('ag-grid Selection State', () => {
 
             describe('Range selection behaviour', () => {
                 test('CTRL-click and CMD-click selects multiple rows', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(2);
                     clickRowByIndex(5, { metaKey: true });
@@ -165,7 +184,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test('Single click after multiple selection clears previous selection', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     selectRowsByIndex([1, 3, 5], api);
 
@@ -175,7 +194,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test('SHIFT-click selects range of rows', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(2);
                     clickRowByIndex(5, { shiftKey: true });
@@ -184,7 +203,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test('SHIFT-click extends range downwards from from last selected row', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     selectRowsByIndex([1, 3], api);
 
@@ -194,7 +213,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test('SHIFT-click extends range upwards from from last selected row', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     selectRowsByIndex([2, 4], api);
 
@@ -204,7 +223,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test('SHIFT-click on un-selected table selects only clicked row', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(4, { shiftKey: true });
                     assertSelectedRowsByIndex([4], api);
@@ -214,7 +233,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test('Range selection is preserved on CTRL-click and CMD-click', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(1);
                     clickRowByIndex(3, { shiftKey: true });
@@ -225,7 +244,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test('Range members can be un-selected with CTRL-click or CMD-click', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(1);
                     clickRowByIndex(4, { shiftKey: true });
@@ -239,7 +258,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test('Range is extended downwards from selection root', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(2);
                     clickRowByIndex(4, { shiftKey: true });
@@ -250,7 +269,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test('Range is extended upwards from selection root', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(6);
                     clickRowByIndex(4, { shiftKey: true });
@@ -261,7 +280,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test('Range can be inverted', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(4);
                     clickRowByIndex(6, { shiftKey: true });
@@ -272,7 +291,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test.skip('SHIFT-click within range after de-selection resets root and clears previous selection', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(2);
                     clickRowByIndex(6, { shiftKey: true });
@@ -286,7 +305,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test.skip('SHIFT-click below range after de-selection resets root and clears previous selection', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(2);
                     clickRowByIndex(5, { shiftKey: true });
@@ -300,7 +319,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test.skip('SHIFT-click above range after de-selection resets root and clears previous selection', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(2);
                     clickRowByIndex(5, { shiftKey: true });
@@ -314,7 +333,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test.skip('META+SHIFT-click within range allows batch deselection', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(2);
                     clickRowByIndex(6, { shiftKey: true });
@@ -328,7 +347,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test.skip('META+SHIFT-click below range allows batch deselection', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(2);
                     clickRowByIndex(5, { shiftKey: true });
@@ -342,7 +361,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test.skip('META+SHIFT-click above range allows batch deselection', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(2);
                     clickRowByIndex(5, { shiftKey: true });
@@ -356,7 +375,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test.skip('CTRL+SHIFT-click within range allows batch deselection', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(2);
                     clickRowByIndex(6, { shiftKey: true });
@@ -370,7 +389,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test.skip('CTRL+SHIFT-click below range allows batch deselection', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(2);
                     clickRowByIndex(5, { shiftKey: true });
@@ -384,7 +403,7 @@ describe('ag-grid Selection State', () => {
                 });
 
                 test.skip('CTRL+SHIFT-click above range allows batch deselection', () => {
-                    const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
+                    const api = createGrid({ columnDefs, rowData, selection: { mode: 'multiRow' } });
 
                     clickRowByIndex(2);
                     clickRowByIndex(5, { shiftKey: true });
@@ -404,8 +423,7 @@ describe('ag-grid Selection State', () => {
                 const api = createGrid({
                     columnDefs,
                     rowData,
-                    rowSelection: 'multiple',
-                    rowMultiSelectWithClick: true,
+                    selection: { mode: 'multiRow', enableMultiSelectWithClick: true },
                 });
 
                 clickRowByIndex(2);
@@ -419,8 +437,7 @@ describe('ag-grid Selection State', () => {
                 const api = createGrid({
                     columnDefs,
                     rowData,
-                    rowSelection: 'multiple',
-                    rowMultiSelectWithClick: true,
+                    selection: { mode: 'multiRow', enableMultiSelectWithClick: true },
                 });
 
                 selectRowsByIndex([1, 2, 3], api);
@@ -431,12 +448,13 @@ describe('ag-grid Selection State', () => {
             });
         });
 
-        describe('Checkbox selection', () => {
-            test('Checkbox can be toggled on and off', () => {
-                const api = createGrid({
-                    columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+        // TODO: Re-enable once checkbox settings are used/ported in the new API
+        describe.skip('Checkbox selection', () => {
+            test('Checkbox can be toggled on and off', async () => {
+                const api = await createGridAndWait({
+                    columnDefs,
                     rowData,
-                    rowSelection: 'multiple',
+                    selection: { mode: 'multiRow', checkboxes: true, suppressClickSelection: true },
                 });
 
                 toggleCheckboxByIndex(1);
@@ -448,9 +466,9 @@ describe('ag-grid Selection State', () => {
 
             test('Multiple rows can be selected without modifier keys nor rowMultiSelectWithClick', () => {
                 const api = createGrid({
-                    columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                    columnDefs,
                     rowData,
-                    rowSelection: 'multiple',
+                    selection: { mode: 'multiRow', checkboxes: true, suppressClickSelection: true },
                 });
 
                 toggleCheckboxByIndex(1);
@@ -460,15 +478,16 @@ describe('ag-grid Selection State', () => {
                 assertSelectedRowsByIndex([1, 2], api);
             });
 
-            test('Clicking a row still selects it when `suppressRowClickSelection` is false', () => {
+            test('Clicking a row still selects it when `suppressClickSelection` is false', () => {
                 const api = createGrid({
-                    columnDefs: columnDefs.map((d) => ({
-                        ...d,
-                        checkboxSelection: true,
-                        showDisabledCheckboxes: true,
-                    })),
+                    columnDefs,
                     rowData,
-                    rowSelection: 'multiple',
+                    selection: {
+                        mode: 'multiRow',
+                        checkboxes: true,
+                        hideDisabledCheckboxes: false,
+                        suppressClickSelection: false,
+                    },
                 });
 
                 // click, not toggle
@@ -480,16 +499,15 @@ describe('ag-grid Selection State', () => {
                 assertSelectedRowsByIndex([], api);
             });
 
-            test('Clicking a row does nothing when `suppressRowClickSelection` is true', () => {
+            test('Clicking a row does nothing when `suppressClickSelection` is true', () => {
                 const api = createGrid({
-                    columnDefs: columnDefs.map((d) => ({
-                        ...d,
-                        checkboxSelection: true,
-                        showDisabledCheckboxes: true,
-                    })),
+                    columnDefs,
                     rowData,
-                    rowSelection: 'multiple',
-                    suppressRowClickSelection: true,
+                    selection: {
+                        mode: 'multiRow',
+                        checkboxes: true,
+                        suppressClickSelection: true,
+                    },
                 });
 
                 // click, not toggle
@@ -499,14 +517,13 @@ describe('ag-grid Selection State', () => {
 
             test('Un-selectable checkboxes cannot be toggled', () => {
                 const api = createGrid({
-                    columnDefs: columnDefs.map((d) => ({
-                        ...d,
-                        checkboxSelection: true,
-                        showDisabledCheckboxes: true,
-                    })),
+                    columnDefs,
                     rowData,
-                    rowSelection: 'multiple',
-                    isRowSelectable: (node) => node.data.sport !== 'golf',
+                    selection: {
+                        mode: 'multiRow',
+                        checkboxes: true,
+                        isRowSelectable: (node) => node.data.sport !== 'golf',
+                    },
                 });
 
                 toggleCheckboxByIndex(4);
@@ -520,9 +537,9 @@ describe('ag-grid Selection State', () => {
             describe('Range selection behaviour', () => {
                 test('CTRL-click and CMD-click does not affect ability to select multiple rows', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -534,9 +551,9 @@ describe('ag-grid Selection State', () => {
 
                 test('SHIFT-click selects range of rows', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -547,9 +564,9 @@ describe('ag-grid Selection State', () => {
 
                 test('SHIFT-click extends range downwards from from last selected row', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     selectRowsByIndex([1, 3], api);
@@ -561,9 +578,9 @@ describe('ag-grid Selection State', () => {
 
                 test('SHIFT-click extends range upwards from from last selected row', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     selectRowsByIndex([2, 4], api);
@@ -575,9 +592,9 @@ describe('ag-grid Selection State', () => {
 
                 test('SHIFT-click on un-selected table selects only clicked row', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(4, { shiftKey: true });
@@ -589,9 +606,9 @@ describe('ag-grid Selection State', () => {
 
                 test('Range selection is preserved on CTRL-click and CMD-click', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(1);
@@ -604,9 +621,9 @@ describe('ag-grid Selection State', () => {
 
                 test('Range selection is preserved on checkbox toggle', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(1);
@@ -619,9 +636,9 @@ describe('ag-grid Selection State', () => {
 
                 test('Range members can be un-selected with CTRL-click or CMD-click', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(1);
@@ -637,9 +654,9 @@ describe('ag-grid Selection State', () => {
 
                 test('Range members can be un-selected with toggle', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(1);
@@ -652,9 +669,9 @@ describe('ag-grid Selection State', () => {
 
                 test('Range is extended downwards from selection root', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -667,9 +684,9 @@ describe('ag-grid Selection State', () => {
 
                 test('Range is extended upwards from selection root', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(6);
@@ -682,9 +699,9 @@ describe('ag-grid Selection State', () => {
 
                 test('Range can be inverted', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(4);
@@ -697,9 +714,9 @@ describe('ag-grid Selection State', () => {
 
                 test('SHIFT can be used for range de-selection (Checkbox selection ONLY)', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(1);
@@ -715,9 +732,9 @@ describe('ag-grid Selection State', () => {
 
                 test.skip('META+SHIFT-click within range allows batch deselection', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -733,9 +750,9 @@ describe('ag-grid Selection State', () => {
 
                 test.skip('META+SHIFT-click below range allows batch deselection', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -751,9 +768,9 @@ describe('ag-grid Selection State', () => {
 
                 test.skip('META+SHIFT-click above range allows batch deselection', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -769,9 +786,9 @@ describe('ag-grid Selection State', () => {
 
                 test.skip('CTRL+SHIFT-click within range allows batch deselection', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -787,9 +804,9 @@ describe('ag-grid Selection State', () => {
 
                 test.skip('CTRL+SHIFT-click below range allows batch deselection', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -805,9 +822,9 @@ describe('ag-grid Selection State', () => {
 
                 test.skip('CTRL+SHIFT-click above range allows batch deselection', () => {
                     const api = createGrid({
-                        columnDefs: columnDefs.map((d) => ({ ...d, checkboxSelection: true })),
+                        columnDefs,
                         rowData,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow', checkboxes: true },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -823,12 +840,13 @@ describe('ag-grid Selection State', () => {
             });
         });
 
-        describe('Header checkbox selection', () => {
+        // TODO: Re-enable once checkbox settings are used/ported in the new API
+        describe.skip('Header checkbox selection', () => {
             test('can be used to select and deselect all rows', () => {
                 const api = createGrid({
-                    columnDefs: columnDefs.map((d) => ({ ...d, headerCheckboxSelection: true })),
+                    columnDefs,
                     rowData,
-                    rowSelection: 'multiple',
+                    selection: { mode: 'multiRow', headerCheckbox: true },
                 });
 
                 toggleHeaderCheckboxByIndex(0);
@@ -840,9 +858,9 @@ describe('ag-grid Selection State', () => {
 
             test('can select multiple pages of data', () => {
                 const api = createGrid({
-                    columnDefs: columnDefs.map((d) => ({ ...d, headerCheckboxSelection: true })),
+                    columnDefs,
                     rowData,
-                    rowSelection: 'multiple',
+                    selection: { mode: 'multiRow', headerCheckbox: true },
                     pagination: true,
                     paginationPageSize: 5,
                 });
@@ -856,13 +874,13 @@ describe('ag-grid Selection State', () => {
 
             test('can select only current page of data', () => {
                 const api = createGrid({
-                    columnDefs: columnDefs.map((d) => ({
-                        ...d,
-                        headerCheckboxSelection: true,
-                        headerCheckboxSelectionCurrentPageOnly: true,
-                    })),
+                    columnDefs,
                     rowData,
-                    rowSelection: 'multiple',
+                    selection: {
+                        mode: 'multiRow',
+                        headerCheckbox: true,
+                        selectAll: 'currentPage',
+                    },
                     pagination: true,
                     paginationPageSize: 5,
                 });
@@ -876,13 +894,13 @@ describe('ag-grid Selection State', () => {
 
             test('can select only filtered data', () => {
                 const api = createGrid({
-                    columnDefs: columnDefs.map((d) => ({
-                        ...d,
-                        headerCheckboxSelection: true,
-                        headerCheckboxSelectionFilteredOnly: true,
-                    })),
+                    columnDefs,
                     rowData,
-                    rowSelection: 'multiple',
+                    selection: {
+                        mode: 'multiRow',
+                        headerCheckbox: true,
+                        selectAll: 'filtered',
+                    },
                     pagination: true,
                     paginationPageSize: 5,
                 });
@@ -899,12 +917,9 @@ describe('ag-grid Selection State', () => {
 
             test('indeterminate selection state transitions to select all', () => {
                 const api = createGrid({
-                    columnDefs: columnDefs.map((d) => ({
-                        ...d,
-                        headerCheckboxSelection: true,
-                    })),
+                    columnDefs,
                     rowData,
-                    rowSelection: 'multiple',
+                    selection: { mode: 'multiRow', headerCheckbox: true },
                 });
 
                 selectRowsByIndex([3], api);
@@ -915,13 +930,13 @@ describe('ag-grid Selection State', () => {
 
             test('un-selectable rows are not part of the selection', () => {
                 const api = createGrid({
-                    columnDefs: columnDefs.map((d) => ({
-                        ...d,
-                        headerCheckboxSelection: true,
-                    })),
+                    columnDefs,
                     rowData,
-                    rowSelection: 'multiple',
-                    isRowSelectable: (node) => node.data.sport !== 'football',
+                    selection: {
+                        mode: 'multiRow',
+                        headerCheckbox: true,
+                        isRowSelectable: (node) => node.data.sport !== 'football',
+                    },
                 });
 
                 toggleHeaderCheckboxByIndex(0);
@@ -929,7 +944,8 @@ describe('ag-grid Selection State', () => {
             });
         });
 
-        describe('Group checkbox selection', () => {
+        // TODO: Re-enable once checkbox settings are used/ported in the new API
+        describe.skip('Group checkbox selection', () => {
             const groupGridOptions: Partial<GridOptions> = {
                 columnDefs: [
                     { field: 'country', rowGroup: true, hide: true },
@@ -953,7 +969,7 @@ describe('ag-grid Selection State', () => {
             test('clicking group row selects only that row', async () => {
                 const api = await createGridAndWait({
                     ...groupGridOptions,
-                    rowSelection: 'multiple',
+                    selection: { mode: 'multiRow' },
                 });
 
                 toggleCheckboxByIndex(0);
@@ -963,8 +979,7 @@ describe('ag-grid Selection State', () => {
             test('clicking group row with `groupSelectsChildren` enabled selects that row and all its children', async () => {
                 const api = await createGridAndWait({
                     ...groupGridOptions,
-                    rowSelection: 'multiple',
-                    groupSelectsChildren: true,
+                    selection: { mode: 'multiRow', groupSelects: 'descendants' },
                 });
 
                 // Group selects children
@@ -988,7 +1003,7 @@ describe('ag-grid Selection State', () => {
                 test('CTRL-click and CMD-click does not affect ability to select multiple rows', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -1001,7 +1016,7 @@ describe('ag-grid Selection State', () => {
                 test('SHIFT-click selects range of rows', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -1013,7 +1028,7 @@ describe('ag-grid Selection State', () => {
                 test('SHIFT-click extends range downwards from from last selected row', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     selectRowsByIndex([1, 3], api);
@@ -1026,7 +1041,7 @@ describe('ag-grid Selection State', () => {
                 test('SHIFT-click extends range upwards from from last selected row', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     selectRowsByIndex([2, 4], api);
@@ -1039,7 +1054,7 @@ describe('ag-grid Selection State', () => {
                 test('SHIFT-click on un-selected table selects only clicked row', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(4, { shiftKey: true });
@@ -1052,7 +1067,7 @@ describe('ag-grid Selection State', () => {
                 test('Range selection is preserved on CTRL-click and CMD-click', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(1);
@@ -1066,7 +1081,7 @@ describe('ag-grid Selection State', () => {
                 test('Range selection is preserved on checkbox toggle', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(1);
@@ -1080,7 +1095,7 @@ describe('ag-grid Selection State', () => {
                 test('Range members can be un-selected with CTRL-click or CMD-click', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(1);
@@ -1097,7 +1112,7 @@ describe('ag-grid Selection State', () => {
                 test('Range members can be un-selected with toggle', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(1);
@@ -1111,7 +1126,7 @@ describe('ag-grid Selection State', () => {
                 test('Range is extended downwards from selection root', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -1125,7 +1140,7 @@ describe('ag-grid Selection State', () => {
                 test('Range is extended upwards from selection root', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(6);
@@ -1139,7 +1154,7 @@ describe('ag-grid Selection State', () => {
                 test('Range can be inverted', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(4);
@@ -1153,7 +1168,7 @@ describe('ag-grid Selection State', () => {
                 test('SHIFT can be used for range de-selection (Checkbox selection ONLY)', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(1);
@@ -1170,7 +1185,7 @@ describe('ag-grid Selection State', () => {
                 test.skip('META+SHIFT-click within range allows batch deselection', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -1187,7 +1202,7 @@ describe('ag-grid Selection State', () => {
                 test.skip('META+SHIFT-click below range allows batch deselection', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -1204,7 +1219,7 @@ describe('ag-grid Selection State', () => {
                 test.skip('META+SHIFT-click above range allows batch deselection', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -1221,7 +1236,7 @@ describe('ag-grid Selection State', () => {
                 test.skip('CTRL+SHIFT-click within range allows batch deselection', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -1238,7 +1253,7 @@ describe('ag-grid Selection State', () => {
                 test.skip('CTRL+SHIFT-click below range allows batch deselection', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -1255,7 +1270,7 @@ describe('ag-grid Selection State', () => {
                 test.skip('CTRL+SHIFT-click above range allows batch deselection', async () => {
                     const api = await createGridAndWait({
                         ...groupGridOptions,
-                        rowSelection: 'multiple',
+                        selection: { mode: 'multiRow' },
                     });
 
                     toggleCheckboxByIndex(2);
@@ -1268,143 +1283,6 @@ describe('ag-grid Selection State', () => {
                     toggleCheckboxByIndex(1, { shiftKey: true, ctrlKey: true });
                     assertSelectedRowsByIndex([4, 5], api);
                 });
-            });
-        });
-    });
-
-    describe('Selection API', () => {
-        describe('setNodesSelected', () => {
-            test('Select single row in single selection mode', () => {
-                const api = createGrid({ columnDefs, rowData, rowSelection: 'single' });
-
-                const nodes = api.getRenderedNodes();
-                const toSelect = [nodes[3]];
-                api.setNodesSelected({ nodes: toSelect, newValue: true });
-
-                const selectedNodes = api.getSelectedNodes();
-
-                expect(selectedNodes).toHaveLength(1);
-                expect(selectedNodes[0]).toBe(toSelect[0]);
-            });
-
-            test('Select single row in multiple selection mode', () => {
-                const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
-
-                const nodes = api.getRenderedNodes();
-                const toSelect = [nodes[3]];
-                api.setNodesSelected({ nodes: toSelect, newValue: true });
-
-                const selectedNodes = api.getSelectedNodes();
-                expect(selectedNodes).toHaveLength(1);
-                expect(selectedNodes[0]).toBe(toSelect[0]);
-            });
-
-            test('Cannot select multiple rows in single selection mode', () => {
-                const api = createGrid({ columnDefs, rowData, rowSelection: 'single' });
-
-                const nodes = api.getRenderedNodes();
-                const toSelect = [nodes[0], nodes[3], nodes[1]];
-                api.setNodesSelected({ nodes: toSelect, newValue: true });
-
-                const selectedNodes = api.getSelectedNodes();
-
-                expect(selectedNodes).toHaveLength(0);
-                expect(consoleWarnSpy).toHaveBeenCalledTimes(1);
-            });
-
-            test('Can select multiple rows in multiple selection mode', () => {
-                const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
-
-                const nodes = api.getRenderedNodes();
-                const toSelect = [nodes[5], nodes[4], nodes[2]];
-                api.setNodesSelected({ nodes: toSelect, newValue: true });
-
-                const selectedNodes = api.getSelectedNodes();
-                expect(selectedNodes).toHaveLength(3);
-            });
-        });
-
-        describe('selectAll', () => {
-            test('Can select all rows in single selection mode', () => {
-                const api = createGrid({ columnDefs, rowData, rowSelection: 'single' });
-
-                api.selectAll();
-
-                expect(api.getSelectedNodes().length).toBe(rowData.length);
-                expect(consoleWarnSpy).not.toHaveBeenCalled();
-
-                api.deselectAll();
-                expect(api.getSelectedNodes().length).toBe(0);
-            });
-
-            test('Can select all rows in multiple selection mode', () => {
-                const api = createGrid({ columnDefs, rowData, rowSelection: 'multiple' });
-
-                api.selectAll();
-
-                expect(api.getSelectedNodes().length).toBe(rowData.length);
-
-                api.deselectAll();
-                expect(api.getSelectedNodes().length).toBe(0);
-            });
-        });
-
-        describe('selectAllOnCurrentPage', () => {
-            test('Can select all rows on current page in single selection mode', () => {
-                const api = createGrid({
-                    columnDefs,
-                    rowData,
-                    rowSelection: 'single',
-                    pagination: true,
-                    paginationPageSize: 5,
-                });
-
-                api.selectAllOnCurrentPage();
-
-                assertSelectedRowsByIndex([0, 1, 2, 3, 4], api);
-                expect(consoleWarnSpy).not.toHaveBeenCalled();
-            });
-
-            test('Can deselect only rows on current page in single selection mode', () => {
-                const api = createGrid({
-                    columnDefs,
-                    rowData,
-                    rowSelection: 'single',
-                    pagination: true,
-                    paginationPageSize: 5,
-                });
-
-                api.selectAll();
-                api.deselectAllOnCurrentPage();
-
-                assertSelectedRowsByIndex([5, 6], api);
-            });
-        });
-
-        describe('selectAllFiltered', () => {
-            test('Can select all filtered rows in single selection mode', () => {
-                const api = createGrid({ columnDefs, rowData, rowSelection: 'single' });
-
-                api.setGridOption('quickFilterText', 'ing');
-
-                api.selectAllFiltered();
-                const selectedNodes = api.getSelectedNodes();
-
-                expect(selectedNodes.length).toBe(2);
-            });
-
-            test('Can deselect filtered rows only in single selection mode', () => {
-                const api = createGrid({ columnDefs, rowData, rowSelection: 'single' });
-
-                api.selectAll();
-
-                api.setGridOption('quickFilterText', 'ing');
-
-                api.deselectAllFiltered();
-
-                api.setGridOption('quickFilterText', '');
-
-                assertSelectedRowsByIndex([0, 1, 2, 3, 4], api);
             });
         });
     });
